backend/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-community
sentence-transformers
faiss-cpu
newspaper3k
lxml_html_clean
